+++ lib/components/tutorial_dialog/_dialogs.dart
@@ -1,3 +1,4 @@
 export 'package:flutter_resteye/components/tutorial_dialog/animation_change_tutorial_dialog.dart';
 export 'package:flutter_resteye/components/tutorial_dialog/animation_tutorial_dialog.dart';
 export 'package:flutter_resteye/components/tutorial_dialog/app_log_tutorial_dialog.dart';
+export 'package:flutter_resteye/components/tutorial_dialog/arrow.dart';
