+++ lib/components/tutorial_dialog/animation_tutorial_dialog.dart
@@ -43,10 +43,7 @@ class AnimationTutorialDialog {
                   children: [
                     Transform.translate(
                       offset: Offset(0, 16.h),
-                      child: CustomPaint(
-                        painter: TrianglePainter(),
-                        child: SizedBox(height: 32.h, width: 32.w),
-                      ),
+                      child: const Arrow(),
                     ),
                     Card(
                       child: Padding(
