+++ lib/components/tutorial_dialog/app_log_tutorial_dialog.dart
@@ -72,10 +72,7 @@ class AppLogTutorialDialog {
                       angle: math.pi / 1,
                       child: Transform.translate(
                         offset: Offset(0, 16.h),
-                        child: CustomPaint(
-                          painter: TrianglePainter(),
-                          child: SizedBox(height: 32.h, width: 32.w),
-                        ),
+                        child: const Arrow(),
                       ),
                     ),
                   ],
